Create empty GLFW window
GLFW is a toolkit like SDL that manages the creation of an OpenGL
context within a window, and also handles the common events. GLFW seems
to be less functional than SDL, but should be adequate for our purpose
now.

We are now drawing an empty window, which will be the template for
further development. Our next step is to be able to create the OpenGL
context and draw a triangle in it. Subsequently, we will attempt to
incorporate the ImGUI toolkit.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>