Start fixing GLFW window code
We have started the process of fixing the GLFW window code so that
it correctly renders a triangle primitive. Once we have done this,
we will be sure that we have created the OpenGL context correctly.
After this, we won't need to be involved with OpenGL; GLFW and
ImGUI will take care of the rest.

As of now, we are only rendering a blank slate gray window. The
triangle will be rendered once we have set up the vertex buffer
and vertex array objects (which we had neglected to do correctly
in the last commit).

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>