Segregate window and shader code
The window and shader code were tightly integrated together in the
earlier commit. We have now segregated the shader and window code, and
will later inject the shader into the window. The shader also requires
a vertex data structure, but that has not yet been implemented. Once the
vertex data structure is in place, we will be able to render the shader
geometries alongside the ImGUI widgets in every frame. This is expected
to be completed within the next two commits.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>